Fix Mali unbalanced reference counting for gbm bo
We observed that destroying a texture bound to a gbm bo was causing the
bo reference count to decrease. However, there was no corresponding
reference count increase when the texture was originally bound.

This was causing gbm_bo_destroy() to be called on bo's that were
already destroyed, causing memory corruption, leading to X crashes.

This was reported to Rockchip who confirmed this behaviour and
provided a fixed Mali version.

This binary comes from zhixiong.lin in the email titled
"RK3288 Xorg/Mali segfaults" on 14/09/2018. The original filename
was libmali.so_rk3288_x11_gbm_arm32_r14p0

https://phabricator.endlessm.com/T22772